backend/app/requirements.txt
Flask
flask-cors
pillow
pytesseract
torch
torchvision
numpy
opencv-python
openai
anthropic
google-genai
huggingface_hub
python-dotenv
easyocr
textblob
psycopg2-binary
SQLAlchemy
Flask-SQLAlchemy
requests
pandas
beautifulsoup4
lxml
scikit-learn
Authlib
Werkzeug
